requirements.txt
playwright
pyperclip
SpeechRecognition
PyAudio
google-genai
